One-shot signals
This adds support for one-shot signals, which are immediately disconnected
after an emit() call. The API has an (optional) bool added to connect().

The reasoning here is that it turns out to be difficult to extricate an
object's signal handler without accidentally removing them all, or possibly
removing newly added signals.